17/01/25 (2)
...

+++ Season2025/src/main/java/frc/robot/math/Configure.java
@@ -1,22 +1,23 @@
 package frc.robot.math;
 
 public class Configure {
-    private static int side;
-    private static int level;
+    
+    private static byte side;
+    private static byte level;
 
-    public static int getSide(){
+    public static byte getSide(){
         return side;
     }
 
-    public static int getLevel(){
+    public static byte getLevel(){
         return level;
     }
 
-    public static void setSide(int Side){
-        side=Side;
+    public static void setSide(byte Side){
+        side = Side;
     }
 
-    public static void setLevel(int Level){
-        level=Level;
+    public static void setLevel(byte Level){
+        level = Level;
     }
 }
